Factor out shared publish/subscribe utils into stream/utils.
